\name{morlec}
\docType{data}
\alias{morlec}
\title{Sample Data from XPPAUTO Bifurcation Plot}
\description{
            This data set comes from the bifurcation plot produced in the
            XPPAUTO tutorial found at:
            http://www.math.pitt.edu/~bard/bardware/tut/start.html
          }
\usage{morlec}
\format{A Data Frame with 5 Variables.}
}
\keyword{datasets}
