\name{xppauto.plot}
\alias{xppauto.plot}
\title{Bifurcation Graph from Auto (XPP)}
\description{
  This function displays a bifurcation plot from data generated by
  AUTO in XPPAUT (http://www.math.pitt.edu/~bard/xpp/xpp.html).
}
\usage{
  xppauto.plot(data, xlim=NULL, ylim=NULL,
               max.step=NULL, min.step=NULL, ...)
}
\arguments{
  \item{data}{filename of data file from XPPAUT.}
  \item{xlim,ylim}{See \link{plot}}
  \item{max.step}{Maximum step size between points}
  \item{min.step}{Minimum step size between points}
  \item{...}{further graphical parameters.}

}
\examples{
data(morlec)
xppauto.plot(morlec, min.step=.1)
}
\keyword{hplot}
